Updating game and video specific settings

+++ locallib.php
@@ -226,6 +226,13 @@ function local_lor_get_types() {
   return $types;
 }
 
+function local_lor_get_type_name_from_id($type_id) {
+  global $DB;
+
+  $name = $DB->get_record_sql('SELECT id, name FROM {lor_type} WHERE id=?', array($type_id));
+  return $name->name;
+}
+
 function local_lor_get_grades() {
   global $DB;
 
@@ -266,14 +273,20 @@ function local_lor_get_related_parameters($id) {
   return "?type=-1$grades_string$categories_string";
 }
 
-function local_lor_update_item($id, $title, $topics, $categories, $grades, $contributors) {
+function local_lor_update_item($id, $title, $topics, $categories, $grades, $contributors, $link, $width, $height, $video_id) {
   global $DB;
 
   // Update lor_content record.
   $content_record = new stdCLass();
   $content_record->id = $id;
   $content_record->title = $title;
+  $content_record->link = $link;
+  $content_record->width = $width;
+  $content_record->height = $height;
   $DB->update_record('lor_content', $content_record);
+
+  // Check if video ID is set and needs to be updated.
+  if (!is_null($video_id))
 
   // Delete all keywords for the item.
   $DB->delete_records('lor_content_keywords', array('content' => $id));
